requirements.txt
# --- Core Django ---
Django==5.1.2
djangorestframework==3.15.2

# --- Testing ---
pytest==8.3.3
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-html==4.1.1
pytest-cov==6.2.1

# --- Datos falsos y factories ---
factory-boy==3.3.1
Faker==28.4.1

# --- Utilidades ---
requests==2.32.3
python-dotenv==1.0.1

# --- Linter y formato opcional (recomendado para proyectos serios) ---
black==24.8.0
flake8==7.1.1
codecov==2.1.13
psycopg[binary]==3.2.3